requirements.txt
boto3
torch
torchvision
Pillow

